+++ zarr/tests/test_core.py
@@ -18,18 +18,7 @@ from zarr.core import Array
 from zarr.errors import PermissionError
 from zarr.compat import PY2
 from zarr.util import buffer_size
-from zarr.indexing import ix_
 from numcodecs import Delta, FixedScaleOffset, Zlib, Blosc, BZ2
-
-
-def oindex(a, selection):
-    """Implementation of orthogonal indexing with slices and ints."""
-    squeeze_axes = tuple([i for i, s in enumerate(selection) if isinstance(s, int)])
-    selection = ix_(*selection)
-    result = a[selection]
-    if squeeze_axes:
-        result = result.squeeze(axis=squeeze_axes)
-    return result
 
 
 # noinspection PyMethodMayBeStatic
@@ -733,595 +722,6 @@ class TestArray(unittest.TestCase):
         eq(0, z.nchunks_initialized)
         z[:] = 42
         eq(10, z.nchunks_initialized)
-
-    def _test_orthogonal_indexing_1d_common(self, a, z, ix):
-        expect = a[ix]
-        actual = z.get_orthogonal_selection(ix)
-        assert_array_equal(expect, actual)
-        actual = z.oindex[ix]
-        assert_array_equal(expect, actual)
-        # for 1d arrays, also available via __getitem__
-        actual = z[ix]
-        assert_array_equal(expect, actual)
-
-    # noinspection PyStatementEffect
-    def test_orthogonal_indexing_1d_bool(self):
-
-        # setup
-        a = np.arange(1050, dtype=int)
-        z = self.create_array(shape=a.shape, chunks=100, dtype=a.dtype)
-        z[:] = a
-
-        np.random.seed(42)
-        # test with different degrees of sparseness
-        for p in 0.5, 0.1, 0.01:
-            ix = np.random.binomial(1, p, size=a.shape[0]).astype(bool)
-            self._test_orthogonal_indexing_1d_common(a, z, ix)
-
-        # test errors
-        with assert_raises(IndexError):
-            z.oindex[np.zeros(50, dtype=bool)]  # too short
-        with assert_raises(IndexError):
-            z.oindex[np.zeros(2000, dtype=bool)]  # too long
-        with assert_raises(IndexError):
-            z.oindex[[[True, False], [False, True]]]  # too many dimensions
-
-    # noinspection PyStatementEffect
-    def test_orthogonal_indexing_1d_int(self):
-
-        # setup
-        a = np.arange(1050, dtype=int)
-        z = self.create_array(shape=a.shape, chunks=100, dtype=a.dtype)
-        z[:] = a
-
-        np.random.seed(42)
-        # test with different degrees of sparseness
-        for p in 0.5, 0.1, 0.01:
-            ix = np.random.choice(a.shape[0], size=int(a.shape[0] * p), replace=True)
-            self._test_orthogonal_indexing_1d_common(a, z, ix)
-            ix.sort()
-            self._test_orthogonal_indexing_1d_common(a, z, ix)
-
-        # test wraparound
-        ix = [0, 3, 10, -23, -12, -1]
-        expect = a[ix]
-        actual = z.oindex[ix]
-        assert_array_equal(expect, actual)
-
-        # test not sorted
-        ix = [3, 105, 23, 127]  # not monotonically increasing
-        expect = a[ix]
-        actual = z.oindex[ix]
-        assert_array_equal(expect, actual)
-
-        # test errors
-        with assert_raises(IndexError):
-            ix = [a.shape[0] + 1]  # out of bounds
-            z.oindex[ix]
-        with assert_raises(IndexError):
-            ix = [-(a.shape[0] + 1)]  # out of bounds
-            z.oindex[ix]
-        with assert_raises(IndexError):
-            ix = [[2, 4], [6, 8]]  # too many dimensions
-            z.oindex[ix]
-
-    def test_orthogonal_indexing_1d_slice_with_step(self):
-
-        # setup
-        a = np.arange(1050, dtype=int)
-        z = self.create_array(shape=a.shape, chunks=100, dtype=a.dtype)
-        z[:] = a
-
-        selections = [
-            slice(0, 1050),
-            slice(0, 1050, 1),
-            slice(0, 1050, 10),
-            slice(0, 1050, 100),
-            slice(0, 1050, 1000),
-            slice(50, 150, 1),
-            slice(50, 150, 10),
-            slice(50, 150, 100),
-        ]
-        for selection in selections:
-            expect = a[selection]
-            actual = z.get_orthogonal_selection(selection)
-            assert_array_equal(expect, actual)
-            actual = z.oindex[selection]
-            assert_array_equal(expect, actual)
-            # for 1d arrays also available via __getitem__
-            actual = z[selection]
-            assert_array_equal(expect, actual)
-
-    def _test_orthogonal_indexing_2d_common(self, a, z, ix0, ix1):
-
-        selections = [
-            # index both axes with array
-            (ix0, ix1),
-            # mixed indexing with array / slice
-            (ix0, slice(1, 5)),
-            (ix0, slice(1, 5, 2)),
-            (slice(250, 350), ix1),
-            (slice(250, 350, 10), ix1),
-            # mixed indexing with array / int
-            (ix0, 4),
-            (42, ix1),
-        ]
-
-        for selection in selections:
-            expect = oindex(a, selection)
-            actual = z.get_orthogonal_selection(selection)
-            assert_array_equal(expect, actual)
-            actual = z.oindex[selection]
-            assert_array_equal(expect, actual)
-
-    def test_orthogonal_indexing_2d_bool(self):
-
-        # setup
-        a = np.arange(10000, dtype=int).reshape(1000, 10)
-        z = self.create_array(shape=a.shape, chunks=(300, 3), dtype=a.dtype)
-        z[:] = a
-
-        np.random.seed(42)
-        # test with different degrees of sparseness
-        for p in 0.5, 0.1, 0.01:
-            ix0 = np.random.binomial(1, p, size=a.shape[0]).astype(bool)
-            ix1 = np.random.binomial(1, 0.5, size=a.shape[1]).astype(bool)
-
-            # main tests
-            self._test_orthogonal_indexing_2d_common(a, z, ix0, ix1)
-
-            # mixed int array / bool array
-            selections = (
-                (ix0, np.nonzero(ix1)[0]),
-                (np.nonzero(ix0)[0], ix1),
-            )
-            for selection in selections:
-                expect = oindex(a, selection)
-                actual = z.oindex[selection]
-                assert_array_equal(expect, actual)
-
-    def test_orthogonal_indexing_2d_int(self):
-
-        # setup
-        a = np.arange(10000, dtype=int).reshape(1000, 10)
-        z = self.create_array(shape=a.shape, chunks=(300, 3), dtype=a.dtype)
-        z[:] = a
-
-        np.random.seed(42)
-        # test with different degrees of sparseness
-        for p in 0.5, 0.1, 0.01:
-            ix0 = np.random.choice(a.shape[0], size=int(a.shape[0] * p), replace=True)
-            ix1 = np.random.choice(a.shape[1], size=int(a.shape[1] * .5), replace=True)
-            self._test_orthogonal_indexing_2d_common(a, z, ix0, ix1)
-            ix0.sort()
-            ix1.sort()
-            self._test_orthogonal_indexing_2d_common(a, z, ix0, ix1)
-
-    def _test_orthogonal_indexing_3d_common(self, a, z, ix0, ix1, ix2):
-
-        selections = [
-            # index all axes with array
-            (ix0, ix1, ix2),
-            # mixed indexing with single array / slices
-            (ix0, slice(15, 25), slice(1, 5)),
-            (slice(50, 70), ix1, slice(1, 5)),
-            (slice(50, 70), slice(15, 25), ix2),
-            (ix0, slice(15, 25, 5), slice(1, 5, 2)),
-            (slice(50, 70, 3), ix1, slice(1, 5, 2)),
-            (slice(50, 70, 3), slice(15, 25, 5), ix2),
-            # mixed indexing with single array / ints
-            (ix0, 42, 4),
-            (84, ix1, 4),
-            (84, 42, ix2),
-            # mixed indexing with single array / slice / int
-            (ix0, slice(15, 25), 4),
-            (42, ix1, slice(1, 5)),
-            (slice(50, 70), 42, ix2),
-            # mixed indexing with two array / slice
-            (ix0, ix1, slice(1, 5)),
-            (slice(50, 70), ix1, ix2),
-            (ix0, slice(15, 25), ix2),
-            # mixed indexing with two array / integer
-            (ix0, ix1, 4),
-            (42, ix1, ix2),
-            (ix0, 42, ix2),
-        ]
-        for selection in selections:
-            expect = oindex(a, selection)
-            actual = z.get_orthogonal_selection(selection)
-            assert_array_equal(expect, actual)
-            actual = z.oindex[selection]
-            assert_array_equal(expect, actual)
-
-    def test_orthogonal_indexing_3d_bool(self):
-
-        # setup
-        a = np.arange(100000, dtype=int).reshape(200, 50, 10)
-        z = self.create_array(shape=a.shape, chunks=(60, 20, 3), dtype=a.dtype)
-        z[:] = a
-
-        np.random.seed(42)
-        # test with different degrees of sparseness
-        for p in 0.5, 0.1, 0.01:
-            ix0 = np.random.binomial(1, p, size=a.shape[0]).astype(bool)
-            ix1 = np.random.binomial(1, .5, size=a.shape[1]).astype(bool)
-            ix2 = np.random.binomial(1, .5, size=a.shape[2]).astype(bool)
-            self._test_orthogonal_indexing_3d_common(a, z, ix0, ix1, ix2)
-
-    def test_orthogonal_indexing_edge_cases(self):
-
-        a = np.arange(6).reshape(1, 2, 3)
-        z = self.create_array(shape=a.shape, chunks=(1, 2, 3), dtype=a.dtype)
-        z[:] = a
-
-        expect = a[ix_([0], range(2), [0, 1, 2])].squeeze(axis=0)
-        actual = z.oindex[0, :, [0, 1, 2]]
-        assert_array_equal(expect, actual)
-
-        expect = a[ix_([0], range(2), [True, True, True])].squeeze(axis=0)
-        actual = z.oindex[0, :, [True, True, True]]
-        assert_array_equal(expect, actual)
-
-    def test_orthogonal_indexing_3d_int(self):
-
-        # setup
-        a = np.arange(100000, dtype=int).reshape(200, 50, 10)
-        z = self.create_array(shape=a.shape, chunks=(60, 20, 3), dtype=a.dtype)
-        z[:] = a
-
-        np.random.seed(42)
-        # test with different degrees of sparseness
-        for p in 0.5, 0.1, 0.01:
-            ix0 = np.random.choice(a.shape[0], size=int(a.shape[0] * p), replace=True)
-            ix1 = np.random.choice(a.shape[1], size=int(a.shape[1] * .5), replace=True)
-            ix2 = np.random.choice(a.shape[2], size=int(a.shape[2] * .5), replace=True)
-            self._test_orthogonal_indexing_3d_common(a, z, ix0, ix1, ix2)
-            ix0.sort()
-            ix1.sort()
-            ix2.sort()
-            self._test_orthogonal_indexing_3d_common(a, z, ix0, ix1, ix2)
-
-    def _test_orthogonal_indexing_1d_common_set(self, v, a, z, ix):
-        a[:] = 0
-        a[ix] = v[ix]
-        z[:] = 0
-        z.oindex[ix] = v[ix]
-        assert_array_equal(a, z[:])
-        z[:] = 0
-        z.set_orthogonal_selection(ix, v[ix])
-        assert_array_equal(a, z[:])
-        # also available via __getitem__ for 1d arrays
-        z[:] = 0
-        z[ix] = v[ix]
-        assert_array_equal(a, z[:])
-
-    def test_orthogonal_indexing_1d_bool_set(self):
-
-        # setup
-        v = np.arange(1050, dtype=int)
-        a = np.empty(v.shape, dtype=int)
-        z = self.create_array(shape=a.shape, chunks=100, dtype=a.dtype)
-
-        np.random.seed(42)
-        # test with different degrees of sparseness
-        for p in 0.5, 0.1, 0.01:
-            ix = np.random.binomial(1, p, size=a.shape[0]).astype(bool)
-            self._test_orthogonal_indexing_1d_common_set(v, a, z, ix)
-
-    def test_orthogonal_indexing_1d_int_set(self):
-
-        # setup
-        v = np.arange(1050, dtype=int)
-        a = np.empty(v.shape, dtype=v.dtype)
-        z = self.create_array(shape=a.shape, chunks=100, dtype=a.dtype)
-
-        np.random.seed(42)
-        # test with different degrees of sparseness
-        for p in 0.5, 0.1, 0.01:
-            ix = np.random.choice(a.shape[0], size=int(a.shape[0] * p), replace=True)
-            self._test_orthogonal_indexing_1d_common_set(v, a, z, ix)
-            ix.sort()
-            self._test_orthogonal_indexing_1d_common_set(v, a, z, ix)
-
-    def _test_orthogonal_indexing_2d_common_set(self, v, a, z, ix0, ix1):
-
-        selections = (
-            # index both axes with array
-            (ix0, ix1),
-            # mixed indexing with array / slice or int
-            (ix0, slice(1, 5)),
-            (slice(250, 350), ix1),
-            (ix0, 4),
-            (42, ix1),
-        )
-        for selection in selections:
-            a[:] = 0
-            a[ix_(*selection)] = v[ix_(*selection)]
-            z[:] = 0
-            z.oindex[selection] = oindex(v, selection)
-            assert_array_equal(a, z[:])
-            z[:] = 0
-            z.set_orthogonal_selection(selection, oindex(v, selection))
-            assert_array_equal(a, z[:])
-
-    def test_orthogonal_indexing_2d_bool_set(self):
-
-        # setup
-        v = np.arange(10000, dtype=int).reshape(1000, 10)
-        a = np.empty_like(v)
-        z = self.create_array(shape=a.shape, chunks=(300, 3), dtype=a.dtype)
-
-        np.random.seed(42)
-        # test with different degrees of sparseness
-        for p in 0.5, 0.1, 0.01:
-            ix0 = np.random.binomial(1, p, size=a.shape[0]).astype(bool)
-            ix1 = np.random.binomial(1, .5, size=a.shape[1]).astype(bool)
-            self._test_orthogonal_indexing_2d_common_set(v, a, z, ix0, ix1)
-
-    def test_orthogonal_indexing_2d_int_set(self):
-
-        # setup
-        v = np.arange(10000, dtype=int).reshape(1000, 10)
-        a = np.empty_like(v)
-        z = self.create_array(shape=a.shape, chunks=(300, 3), dtype=a.dtype)
-
-        np.random.seed(42)
-        # test with different degrees of sparseness
-        for p in 0.5, 0.1, 0.01:
-            ix0 = np.random.choice(a.shape[0], size=int(a.shape[0] * p), replace=True)
-            ix1 = np.random.choice(a.shape[1], size=int(a.shape[1] * .5), replace=True)
-            self._test_orthogonal_indexing_2d_common_set(v, a, z, ix0, ix1)
-            ix0.sort()
-            ix1.sort()
-            self._test_orthogonal_indexing_2d_common_set(v, a, z, ix0, ix1)
-
-    def _test_orthogonal_indexing_3d_common_set(self, v, a, z, ix0, ix1, ix2):
-
-        selections = (
-            # index all axes with bool array
-            (ix0, ix1, ix2),
-            # mixed indexing with single bool array / slice or int
-            (ix0, slice(15, 25), slice(1, 5)),
-            (slice(50, 70), ix1, slice(1, 5)),
-            (slice(50, 70), slice(15, 25), ix2),
-            (ix0, 42, 4),
-            (84, ix1, 4),
-            (84, 42, ix2),
-            (ix0, slice(15, 25), 4),
-            (slice(50, 70), ix1, 4),
-            (slice(50, 70), 42, ix2),
-            # indexing with two arrays / slice
-            (ix0, ix1, slice(1, 5)),
-            # indexing with two arrays / integer
-            (ix0, ix1, 4),
-        )
-        for selection in selections:
-            a[:] = 0
-            a[ix_(*selection)] = v[ix_(*selection)]
-            z[:] = 0
-            z.oindex[selection] = oindex(v, selection)
-            assert_array_equal(a, z[:])
-            z[:] = 0
-            z.set_orthogonal_selection(selection, oindex(v, selection))
-            assert_array_equal(a, z[:])
-
-    def test_orthogonal_indexing_3d_bool_set(self):
-
-        # setup
-        v = np.arange(100000, dtype=int).reshape(200, 50, 10)
-        a = np.empty_like(v)
-        z = self.create_array(shape=a.shape, chunks=(60, 20, 3), dtype=a.dtype)
-
-        np.random.seed(42)
-        # test with different degrees of sparseness
-        for p in 0.5, 0.1, 0.01:
-            ix0 = np.random.binomial(1, p, size=a.shape[0]).astype(bool)
-            ix1 = np.random.binomial(1, .5, size=a.shape[1]).astype(bool)
-            ix2 = np.random.binomial(1, .5, size=a.shape[2]).astype(bool)
-            self._test_orthogonal_indexing_3d_common_set(v, a, z, ix0, ix1, ix2)
-
-    def test_orthogonal_indexing_3d_int_set(self):
-
-        # setup
-        v = np.arange(100000, dtype=int).reshape(200, 50, 10)
-        a = np.empty_like(v)
-        z = self.create_array(shape=a.shape, chunks=(60, 20, 3), dtype=a.dtype)
-
-        np.random.seed(42)
-        # test with different degrees of sparseness
-        for p in 0.5, 0.1, 0.01:
-            ix0 = np.random.choice(a.shape[0], size=int(a.shape[0] * p), replace=True)
-            ix1 = np.random.choice(a.shape[1], size=int(a.shape[1] * .5), replace=True)
-            ix2 = np.random.choice(a.shape[2], size=int(a.shape[2] * .5), replace=True)
-            self._test_orthogonal_indexing_3d_common_set(v, a, z, ix0, ix1, ix2)
-            ix0.sort()
-            ix1.sort()
-            ix2.sort()
-            self._test_orthogonal_indexing_3d_common_set(v, a, z, ix0, ix1, ix2)
-
-    # noinspection PyStatementEffect
-    def test_coordinate_indexing_1d(self):
-
-        # setup
-        a = np.arange(1050, dtype=int)
-        z = self.create_array(shape=a.shape, chunks=100, dtype=a.dtype)
-        z[:] = a
-
-        np.random.seed(42)
-        # test with different degrees of sparseness
-        for p in 0.5, 0.1, 0.01:
-            ix = np.random.choice(a.shape[0], size=int(a.shape[0] * p), replace=True)
-            expect = a[ix]
-            actual = z.get_coordinate_selection(ix)
-            assert_array_equal(expect, actual)
-            actual = z.vindex[ix]
-            assert_array_equal(expect, actual)
-            ix.sort()
-            expect = a[ix]
-            actual = z.get_coordinate_selection(ix)
-            assert_array_equal(expect, actual)
-            actual = z.vindex[ix]
-            assert_array_equal(expect, actual)
-
-        # test single item
-        ix = 42
-        expect = a[ix]
-        actual = z.get_coordinate_selection(ix)
-        assert_array_equal(expect, actual)
-
-        # test wraparound
-        ix = [0, 3, 10, -23, -12, -1]
-        expect = a[ix]
-        actual = z.get_coordinate_selection(ix)
-        assert_array_equal(expect, actual)
-
-        # test out of order
-        ix = [3, 105, 23, 127]  # not monotonically increasing
-        expect = a[ix]
-        actual = z.get_coordinate_selection(ix)
-        assert_array_equal(expect, actual)
-
-        # test errors
-        with assert_raises(IndexError):
-            ix = [a.shape[0] + 1]  # out of bounds
-            z.get_coordinate_selection(ix)
-        with assert_raises(IndexError):
-            ix = [-(a.shape[0] + 1)]  # out of bounds
-            z.get_coordinate_selection(ix)
-        with assert_raises(IndexError):
-            ix = [[2, 4], [6, 8]]  # too many dimensions
-            z.get_coordinate_selection(ix)
-        with assert_raises(IndexError):
-            ix = slice(5, 15)
-            z.get_coordinate_selection(ix)
-        with assert_raises(IndexError):
-            ix = Ellipsis
-            z.get_coordinate_selection(ix)
-
-    def test_coordinate_indexing_2d(self):
-
-        # setup
-        a = np.arange(10000, dtype=int).reshape(1000, 10)
-        z = self.create_array(shape=a.shape, chunks=(300, 3), dtype=a.dtype)
-        z[:] = a
-
-        np.random.seed(42)
-        # test with different degrees of sparseness
-        for p in 0.5, 0.1, 0.01:
-            n = int(a.size * p)
-            ix0 = np.random.choice(a.shape[0], size=n, replace=True)
-            ix1 = np.random.choice(a.shape[1], size=n, replace=True)
-            selections = [
-                # index both axes with array
-                (ix0, ix1),
-                # mixed indexing with array / int
-                (ix0, 4),
-                (42, ix1),
-                (42, 4),
-            ]
-
-            for selection in selections:
-                expect = a[selection]
-                actual = z.get_coordinate_selection(selection)
-                assert_array_equal(expect, actual)
-                actual = z.vindex[selection]
-                assert_array_equal(expect, actual)
-
-            srt = np.lexsort((ix0, ix1))
-            ix0 = ix0[srt]
-            ix1 = ix1[srt]
-            selections = [
-                # index both axes with array
-                (ix0, ix1),
-                # mixed indexing with array / int
-                (ix0, 4),
-                (42, ix1),
-                (42, 4),
-            ]
-
-            for selection in selections:
-                expect = a[selection]
-                actual = z.get_coordinate_selection(selection)
-                assert_array_equal(expect, actual)
-                actual = z.vindex[selection]
-                assert_array_equal(expect, actual)
-
-        # not monotonically increasing (first dim)
-        ix0 = [3, 3, 4, 2, 5]
-        ix1 = [1, 3, 5, 7, 9]
-        expect = a[ix0, ix1]
-        actual = z.get_coordinate_selection((ix0, ix1))
-        assert_array_equal(expect, actual)
-
-        # not monotonically increasing (second dim)
-        ix0 = [1, 1, 2, 2, 5]
-        ix1 = [1, 3, 2, 1, 0]
-        expect = a[ix0, ix1]
-        actual = z.get_coordinate_selection((ix0, ix1))
-        assert_array_equal(expect, actual)
-
-        with assert_raises(IndexError):
-            selection = slice(5, 15), [1, 2, 3]
-            z.get_coordinate_selection(selection)
-        with assert_raises(IndexError):
-            selection = [1, 2, 3], slice(5, 15)
-            z.get_coordinate_selection(selection)
-        with assert_raises(IndexError):
-            selection = Ellipsis, [1, 2, 3]
-            z.get_coordinate_selection(selection)
-
-    def test_get_selection_out(self):
-
-        # basic selections
-        a = np.arange(1050)
-        z = self.create_array(shape=1050, chunks=100, dtype=a.dtype)
-        z[:] = a
-        selections = [
-            slice(50, 150),
-            slice(0, 1050),
-            slice(1, 2),
-        ]
-        for selection in selections:
-            expect = a[selection]
-            out = self.create_array(shape=expect.shape, chunks=10, dtype=expect.dtype, fill_value=0)
-            z.get_basic_selection(selection, out=out)
-            assert_array_equal(expect, out[:])
-
-        # orthogonal selections
-        a = np.arange(10000, dtype=int).reshape(1000, 10)
-        z = self.create_array(shape=a.shape, chunks=(300, 3), dtype=a.dtype)
-        z[:] = a
-        np.random.seed(42)
-        # test with different degrees of sparseness
-        for p in 0.5, 0.1, 0.01:
-            ix0 = np.random.binomial(1, p, size=a.shape[0]).astype(bool)
-            ix1 = np.random.binomial(1, .5, size=a.shape[1]).astype(bool)
-            selections = [
-                # index both axes with array
-                (ix0, ix1),
-                # mixed indexing with array / slice
-                (ix0, slice(1, 5)),
-                (slice(250, 350), ix1),
-                # mixed indexing with array / int
-                (ix0, 4),
-                (42, ix1),
-                # mixed int array / bool array
-                (ix0, np.nonzero(ix1)[0]),
-                (np.nonzero(ix0)[0], ix1),
-            ]
-            for selection in selections:
-                expect = oindex(a, selection)
-                # out = self.create_array(shape=expect.shape, chunks=10, dtype=expect.dtype,
-                #                         fill_value=0)
-                out = np.zeros(expect.shape, dtype=expect.dtype)
-                z.get_orthogonal_selection(selection, out=out)
-                assert_array_equal(expect, out[:])
-
-        # TODO coordinate selection
-
-        # TODO mask selection
-
-        # TODO selection with fields
 
 
 class TestArrayWithPath(TestArray):
